Enhance API configuration and card normalization logic
- Added BASE_URL and SWAGGER_BASE_URL to .env.example for improved API configuration.
- Updated CLAUDE.md to document new environment variables for Swagger/OpenAPI servers.
- Refactored API server configuration to utilize BASE_URL for Swagger UI.
- Implemented a new normalize.ts file for shared card name normalization logic, improving consistency across providers.
- Updated RiftCodex and Supabase providers to use the new normalization function.
