+++ .env.example
@@ -25,6 +25,10 @@ FUZZY_THRESHOLD=0.4
 # ─── Elysia API ────────────────────────────────────────────────────────────────
 API_PORT=3000
 
+# Base URL for Swagger UI / OpenAPI servers (e.g. behind reverse proxy). Fallback: /
+# BASE_URL=/
+# SWAGGER_BASE_URL=https://api.example.com
+
 # Public-facing URL of this API (used in bot reply links)
 API_BASE_URL=http://localhost:3000
